# swagger/swagger.yaml
definitions:
  github_com_yourusername_subscription-service_internal_model.Subscription:
    properties:
      end_date:
        description: Опционально
        type: string
      id:
        description: UUID
        type: string
      price:
        description: Стоимость в рублях
        type: integer
      service_name:
        description: Название сервиса
        type: string
      start_date:
        description: Месяц и год начала
        type: string
      user_id:
        description: UUID пользователя
        type: string
    type: object
info:
  contact: {}
paths:
  /subscriptions:
    get:
      consumes:
      - application/json
      description: Get all subscriptions with optional filtering
      parameters:
      - description: User ID
        in: query
        name: user_id
        type: string
      - description: Service Name
        in: query
        name: service_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_yourusername_subscription-service_internal_model.Subscription'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Create a new subscription
      parameters:
      - description: Subscription info
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/github_com_yourusername_subscription-service_internal_model.Subscription'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_yourusername_subscription-service_internal_model.Subscription'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create subscription
      tags:
      - subscriptions
  /subscriptions/summary:
    get:
      consumes:
      - application/json
      description: Get total subscription cost for a period
      parameters:
      - description: User ID
        in: query
        name: user_id
        type: string
      - description: Service Name
        in: query
        name: service_name
        type: string
      - description: Start month YYYY-MM
        in: query
        name: from
        required: true
        type: string
      - description: End month YYYY-MM
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Sum subscriptions
      tags:
      - subscriptions
swagger: "2.0"